refactor(crashlytics): remove deprecated namespaced API
