+++ packages/app/lib/internal/constants.ts
@@ -23,7 +23,6 @@ export const KNOWN_NAMESPACES = [
   'auth',
   'analytics',
   'remoteConfig',
-  'crashlytics',
   'database',
   'firestore',
   'indexing',
