+++ packages/crashlytics/typedoc.json
@@ -1,6 +1,5 @@
 {
   "$schema": "https://typedoc.org/schema.json",
-  "entryPoints": ["lib/modular.ts", "lib/types/crashlytics.ts"],
-  "tsconfig": "tsconfig.json",
-  "intentionallyNotExported": ["_Statics"]
+  "entryPoints": ["lib/index.ts", "lib/types/crashlytics.ts"],
+  "tsconfig": "tsconfig.json"
 }
